Boys Room Update
I gave all the objects in the boys room a collider (some have a mesh collider and it messes Gooi up when he touches it. Hoping to figure out that bug soon). I have also added Gooi to the boys room and I made a script for the camera that follows gooi the correct way (not just a child of Gooi).