collectible pick up
I created a simple object to test the collectibles. I then made an empty object with a box collider on it big enough for Gooi to fit into. Then I wrote some code that would only allow you to pick up the object (with space) when you are in the box. Still trying to figure out how to make a little text box pop up when you are able to pick up the object.